Use uintptr_t for the pointer<->integer round-trips, for 32-bit builds
func_mod_square()'s fwd_fft_only parameter is a uint64 carrying a double* with
mode flags packed into its low bits: callers hand it (uint64)ptr + mode_flag,
and the radix{16,32}_{dyadic,wrapper}_square routines take the pointer back out
with (double *)(fwd_fft_only & ~0xCull). On a 64-bit build those casts are the
same width and nothing is said; on 32-bit each one is a cast between a 4-byte
pointer and an 8-byte integer, which is what -Wpointer-to-int-cast and
-Wint-to-pointer-cast are reporting.

The value round-trips either way, so this is a warning-only change - but it is
worth making, because the same files already use uintptr_t for exactly this:
radix16_dyadic_square.c:283 tests ((uintptr_t)sm_ptr & 0x3f) a dozen lines from
a cast that did not. This brings the stragglers in line with that.

Three shapes are covered:

  - a pointer packed into the uint64 argument, and unpacked from it, now goes
    through uintptr_t at both ends;
  - the alignment tests in factor.c and mi64.c use (uintptr_t) rather than
    (uint64), since the width of a pointer is what they are asking about;
  - util.c's pthread_join status is a void*, so it is read through intptr_t
    instead of being cast straight to int64.

Left alone deliberately: pm1.c:514 casts the int that fgetc() returns, not a
pointer, and pm1.c:1613 is commented-out code.

Verified on i686 under Debian 13: the cast warnings go from 18 to 0, and -s tt
produces identical Res64 at all 44 radix sets before and after. On x86_64 the
build is warning-free and the same 44 Res64 values match main exactly.

+++ src/Mlucas.c
@@ -1968,7 +1968,7 @@ READ_RESTART_FILE:
 			step and force the undo-initial-FFT-pass-and-DWT-weighting step, leaving a pure-int G-check residue ready for savefile-writing: */
 				mode_flag = 3 - first_sub - (last_sub<<1);
 			//	printf("Iter %u: FFT(b)*FFT(c) step.\n",i);
-				ierr = func_mod_square  (b, (int*)arrtmp, n, i,i+1, (uint64)c + (uint64)mode_flag, p, scrnFlag, &tdif2, FALSE, 0x0);
+				ierr = func_mod_square  (b, (int*)arrtmp, n, i,i+1, (uint64)(uintptr_t)c + (uint64)mode_flag, p, scrnFlag, &tdif2, FALSE, 0x0);
 				if(ierr) {
 					if(ierr == ERR_INTERRUPT) {
 						fprintf(stderr,"Caught interrupt in FFT(b)*FFT(c) step.\n");
